504: added shadow and light mode to tooltip
