+++ src/components/SwapWidget/subcomponents/SwapInputs/SwapInputs.tsx
@@ -7,7 +7,7 @@ import stringToSignificantDecimals from "../../../../helpers/stringToSignificant
 import TokenSelect from "../../../TokenSelect/TokenSelect";
 import getSwapInputIcon from "../../helpers/getSwapInputIcon";
 import getTokenMaxInfoText from "../../helpers/getTokenMaxInfoText";
-import ToolTip from "../ToolTip/ToolTip";
+import Tooltip from "../Tooltip/Tooltip";
 import { Container, SwapIconContainer } from "./SwapInputs.styles";
 
 const floatRegExp = new RegExp("^([0-9])*[.,]?([0-9])*$");
@@ -123,7 +123,7 @@ const SwapInputs: FC<{
       />
       {(showMaxButton || showMaxInfoButton) &&
         showMaxAmountInfo &&
-        maxAmountInfoText && <ToolTip>{maxAmountInfoText}</ToolTip>}
+        maxAmountInfoText && <Tooltip>{maxAmountInfoText}</Tooltip>}
     </Container>
   );
 };
