+++ src/components/SwapWidget/subcomponents/Tooltip/Tooltip.styles.tsx
@@ -30,7 +30,9 @@ export const ContentContainer = styled.div`
   line-height: 1.2;
   font-size: 0.8125rem;
   color: ${(props) => props.theme.colors.white};
-  background: ${(props) => props.theme.colors.darkGrey};
+  background: ${({ theme }) =>
+    theme.name === "dark" ? theme.colors.darkGrey : theme.colors.primaryLight};
+  filter: drop-shadow(${(props) => props.theme.shadows.tooltipGlow});
   pointer-events: none;
 `;
 
@@ -41,29 +43,15 @@ export const Triangle = styled.div`
   height: ${triangleSize}px;
   bottom: -${triangleSize / 2}px;
   left: calc(100% - 2rem);
-  overflow: hidden;
-
-  &:after {
-    content: "";
-    display: block;
-    position: absolute;
-    top: -1px;
-    left: 1px;
-    width: ${triangleSize}px;
-    height: ${triangleSize}px;
-    background: ${(props) => props.theme.colors.darkGrey};
-  }
+  border: 1px solid ${(props) => props.theme.colors.borderGrey};
+  border-top: 0;
+  border-right: 0;
+  background: ${({ theme }) =>
+    theme.name === "dark" ? theme.colors.darkGrey : theme.colors.primaryLight};
 
   @media ${breakPoints.tabletLandscapeUp} {
     transform: rotate(45deg);
     top: calc(50% - ${triangleSize / 2}px);
     left: -${triangleSize / 2}px;
   }
-`;
-
-export const Square = styled.div`
-  width: ${triangleSize}px;
-  height: ${triangleSize}px;
-  border: 1px solid ${(props) => props.theme.colors.borderGrey};
-  background: ${(props) => props.theme.colors.darkGrey};
 `;
